Fix authorization error handling

+++ internal/controllers/comments.go
@@ -2,6 +2,7 @@ package controllers
 
 import (
 	errs "cvwo-backend/internal/errors"
+	"cvwo-backend/internal/middleware"
 	"cvwo-backend/internal/models"
 	"cvwo-backend/internal/services"
 	"net/http"
@@ -22,7 +23,7 @@ func NewCommentController(service services.CommentService) *CommentController {
 func (controller *CommentController) GetByPostID(ctx *gin.Context) {
 	postId, err := strconv.Atoi(ctx.Param("id"))
 	if err != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": "invalid post ID"})
+		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Invalid post ID"})
 		return
 	}
 
@@ -45,19 +46,17 @@ func (controller *CommentController) Create(ctx *gin.Context) {
 	}
 
 	// Retrieve the authenticated user from context
-	// user, exists := ctx.Get("user")
-	// // This should not happen as middleware already checks for valid user
-	// if !exists {
-	// 	ctx.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
-	// 	return
-	// }
+	user, err := middleware.GetUserFromContext(ctx)
+	if err != nil {
+		errs.HTTPErrorResponse(ctx, err)
+		return
+	}
 
-	// // Check that the authorID of the comment corresponds to the currently authenticated user's ID
-	// userID := user.(*models.User).ID
-	// if userID != requestBody.AuthorID {
-	// 	ctx.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
-	// 	return
-	// }
+	// Check that the authorID of the comment corresponds to the currently authenticated user's ID
+	if user.ID != requestBody.AuthorID {
+		ctx.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
+		return
+	}
 
 	// Map fields from request body to Comment model
 	comment := models.Comment{
@@ -66,9 +65,8 @@ func (controller *CommentController) Create(ctx *gin.Context) {
 		AuthorID: requestBody.AuthorID,
 	}
 
+	// Create new comment
 	newComment, err := controller.service.Create(&comment)
-
-	// Handle errors
 	if err != nil {
 		errs.HTTPErrorResponse(ctx, err)
 		return
@@ -82,24 +80,29 @@ func (controller *CommentController) Update(ctx *gin.Context) {
 	// Validate commentID
 	id, err := strconv.Atoi(ctx.Param("id"))
 	if err != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": "invalid comment ID"})
+		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Invalid comment ID"})
 		return
 	}
 
-	// // Retrieve the authenticated user from context
-	// user, exists := ctx.Get("user")
-	// // This should not happen as middleware already checks for valid user
-	// if !exists {
-	// 	ctx.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
-	// 	return
-	// }
+	// Retrieve the authenticated user from context
+	user, err := middleware.GetUserFromContext(ctx)
+	if err != nil {
+		errs.HTTPErrorResponse(ctx, err)
+		return
+	}
 
-	// // TODO: Check that the authorID of the comment corresponds to the currently authenticated user's ID
-	// userID := user.(*models.User).ID
-	// if userID != uint(id) {
-	// 	ctx.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
-	// 	return
-	// }
+	// Fetch the comment to check its authorID
+	comment, err := controller.service.GetByID(uint(id))
+	if err != nil {
+		errs.HTTPErrorResponse(ctx, err)
+		return
+	}
+
+	// Check that the authorID of the comment corresponds to the currently authenticated user's ID
+	if user.ID != comment.AuthorID {
+		ctx.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
+		return
+	}
 
 	// Validate request body
 	var requestBody models.CommentUpdate
@@ -108,9 +111,8 @@ func (controller *CommentController) Update(ctx *gin.Context) {
 		return
 	}
 
+	// Update the comment
 	updatedComment, err := controller.service.Update(uint(id), requestBody.Content)
-
-	// Handle errors
 	if err != nil {
 		errs.HTTPErrorResponse(ctx, err)
 		return
@@ -124,25 +126,31 @@ func (controller *CommentController) Delete(ctx *gin.Context) {
 	// Validate commentID
 	id, err := strconv.Atoi(ctx.Param("id"))
 	if err != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": "invalid comment ID"})
+		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Invalid comment ID"})
 		return
 	}
 
 	// Retrieve the authenticated user from context
-	// user, exists := ctx.Get("user")
-	// // This should not happen as middleware already checks for valid user
-	// if !exists {
-	// 	ctx.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
-	// 	return
-	// }
+	user, err := middleware.GetUserFromContext(ctx)
+	if err != nil {
+		errs.HTTPErrorResponse(ctx, err)
+		return
+	}
 
-	// // Check that the authorID of the comment corresponds to the currently authenticated user's ID
-	// userID := user.(*models.User).ID
-	// if userID != uint(id) { // TODO: Compare with authorID of comment, not comment ID!
-	// 	ctx.JSON(http.StatusUnauthorized, gin.H{"error": "unauthorized"})
-	// 	return
-	// }
+	// Fetch the comment to check its authorID
+	comment, err := controller.service.GetByID(uint(id))
+	if err != nil {
+		errs.HTTPErrorResponse(ctx, err)
+		return
+	}
 
+	// Check that the authorID of the comment corresponds to the currently authenticated user's ID
+	if user.ID != comment.AuthorID { 
+		ctx.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
+		return
+	}
+
+	// Delete the comment
 	if err := controller.service.Delete(uint(id)); err != nil {
 		errs.HTTPErrorResponse(ctx, err)
 		return
